chore: fix hand-written clippy warnings toward a clean -D warnings run
CI runs `cargo clippy --workspace --all-targets --all-features -- -D warnings`
(currently non-blocking, pending a cleanup pass). Fix the warnings that live
in hand-written sources; the remaining warnings are in generated codegen and
must be addressed at the trestle templates, not by hand.

- credential_vending: group digit literals (3_590_000 / 3_610_000)
- acceptance/execution: take &Path instead of &PathBuf in journey-state I/O
- acceptance/reporting: replace unwrap-after-is_some with a match guard
- acceptance/tests/journeys: collapse nested if into a let-chain
- acceptance/execution/helpers: demote a stray doc comment to a line comment
- datafusion/tests/managed_table: drop a redundant `as u64` cast
- server/api/staging_tables: drop redundant ..Default on single-field structs
- common/python: hoist nested format! out of the outer format! args

No behavior change; affected tests still pass.

AI-assisted-by: Isaac

+++ crates/datafusion/tests/managed_table.rs
@@ -163,11 +163,7 @@ async fn create_managed_table_round_trip() {
     let snapshot =
         build_catalog_managed_snapshot(engine.as_ref(), &location, &commits, latest as i64, None)
             .expect("build_catalog_managed_snapshot failed");
-    assert_eq!(
-        snapshot.version(),
-        latest as u64,
-        "snapshot at catalog version"
-    );
+    assert_eq!(snapshot.version(), latest, "snapshot at catalog version");
 
     let provider = DeltaScanNext::builder()
         .with_snapshot(Arc::new(snapshot))
